* 'develope' of github.com:0AndWild/Hanghae99-Dogfaw:
  임포트 최적화
  알림목록 단일 삭제, 전체 삭제 예외처리
  no message
  Notification Service(추가) - subscribe쪽에 에러가 계속 떠서 trycatch 구문 적용

+++ src/main/java/com/project/dogfaw/sse/service/NotificationService.java
@@ -9,6 +9,7 @@ package com.project.dogfaw.sse.service;
 
 import com.project.dogfaw.common.exception.CustomException;
 import com.project.dogfaw.common.exception.ErrorCode;
+import com.project.dogfaw.common.exception.StatusResponseDto;
 import com.project.dogfaw.sse.dto.NotificationCountDto;
 import com.project.dogfaw.sse.dto.NotificationDto;
 import com.project.dogfaw.sse.model.Notification;
@@ -18,6 +19,8 @@ import com.project.dogfaw.sse.repository.EmitterRepositoryImpl;
 import com.project.dogfaw.sse.repository.NotificationRepository;
 import com.project.dogfaw.user.model.User;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
 import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
@@ -37,28 +40,32 @@ public class NotificationService {
     private final NotificationRepository notificationRepository;
 
     public SseEmitter subscribe(Long userId, String lastEventId) {
-        //emitter 하나하나 에 고유의 값을 주기 위해
-        String emitterId = makeTimeIncludeId(userId);
+        try {
+            //emitter 하나하나 에 고유의 값을 주기 위해
+            String emitterId = makeTimeIncludeId(userId);
 
-        Long timeout = 60L * 1000L * 60L; // 1시간
-        // 생성된 emiiterId를 기반으로 emitter를 저장
-        SseEmitter emitter = emitterRepository.save(emitterId, new SseEmitter(timeout));
+            Long timeout = 60L * 1000L * 60L; // 1시간
+            // 생성된 emiiterId를 기반으로 emitter를 저장
+            SseEmitter emitter = emitterRepository.save(emitterId, new SseEmitter(timeout));
 
-        //emitter의 시간이 만료된 후 레포에서 삭제
-        emitter.onCompletion(() -> emitterRepository.deleteById(emitterId));
-        emitter.onTimeout(() -> emitterRepository.deleteById(emitterId));
+            //emitter의 시간이 만료된 후 레포에서 삭제
+            emitter.onCompletion(() -> emitterRepository.deleteById(emitterId));
+            emitter.onTimeout(() -> emitterRepository.deleteById(emitterId));
 
-        // 503 에러를 방지하기 위해 처음 연결 진행 시 더미 데이터를 전달
-        String eventId = makeTimeIncludeId(userId);
-        // 수 많은 이벤트 들을 구분하기 위해 이벤트 ID에 시간을 통해 구분을 해줌
-        sendNotification(emitter, eventId, emitterId, "EventStream Created. [userId=" + userId + "]");
+            // 503 에러를 방지하기 위해 처음 연결 진행 시 더미 데이터를 전달
+            String eventId = makeTimeIncludeId(userId);
+            // 수 많은 이벤트 들을 구분하기 위해 이벤트 ID에 시간을 통해 구분을 해줌
+            sendNotification(emitter, eventId, emitterId, "EventStream Created. [userId=" + userId + "]");
 
-        // 클라이언트가 미수신한 Event 목록이 존재할 경우 전송하여 Event 유실을 예방
-        if (hasLostData(lastEventId)) {
-            sendLostData(lastEventId, userId, emitterId, emitter);
+            // 클라이언트가 미수신한 Event 목록이 존재할 경우 전송하여 Event 유실을 예방
+            if (hasLostData(lastEventId)) {
+                sendLostData(lastEventId, userId, emitterId, emitter);
+            }
+            return emitter;
+        }catch (Exception e){
+            throw new CustomException(ErrorCode.FAIL_SUBSCRIBE);
         }
 
-        return emitter;
     }
 
 
@@ -156,13 +163,26 @@ public class NotificationService {
     }
 
     @Transactional
-    public void deleteAllByNotifications(User user) {
+    public ResponseEntity<Object> deleteAllByNotifications(User user) {
         Long receiverId = user.getId();
-        notificationRepository.deleteAllByReceiverId(receiverId);
+        Optional<Notification> existNotification = notificationRepository.findByReceiver(user);
+       if (existNotification.isPresent()){
+           notificationRepository.deleteAllByReceiverId(receiverId);
+           return new ResponseEntity<>(new StatusResponseDto("알림 목록 전체삭제 성공", true), HttpStatus.OK);
+       }else {
+           return new ResponseEntity<>(new StatusResponseDto("삭제할 알림이 존재하지 않습니다",false ), HttpStatus.BAD_REQUEST);
+       }
 
     }
     @Transactional
-    public void deleteByNotifications(Long notificationId) {
-        notificationRepository.deleteById(notificationId);
+    public ResponseEntity<Object> deleteByNotifications(Long notificationId) {
+        Optional<Notification> notification = notificationRepository.findById(notificationId);
+        if(notification.isPresent()){
+            notificationRepository.deleteById(notificationId);
+            return new ResponseEntity<>(new StatusResponseDto("알림 삭제 완료", true), HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(new StatusResponseDto("존재하지 않는 알림입니다",false ), HttpStatus.BAD_REQUEST);
+        }
+
     }
 }
